fix(workflow): prevent job failure if no changes in the build.
When the build that is made in the workflow matches the current build
file (dist/index.js) then there is nothing to commit and the step
was failing, making the entire job to fail.

Add a conditional to only make the changes if there are changes, this
way only if the commit fails the job will be set a failed.